// app/models/data-batch-details.model.ts
export interface DataBatchDetails{
    SOURCE_ID: string;
    SOURCE_NAME: string;
    LOAD_YEAR: number;
    LOAD_PERIOD: number;
    BATCH_ID: number;
    PUBLISH_STATUS: string;
    PUBLISH_ACTION: string;
    CUST_COUNT: number;
    SUPP_COUNT: number;
    TRANS_COUNT: number;
    LINE_COUNT: number;
    TRANS_MIN_DATE: Date;
    TRANS_MAX_DATE: Date;
    AVG_TRANS_PER_DOC: number;
    AVG_TRANS_PER_SUPP: number;
    AVG_TRANS_PER_CUST: number;
    MAX_TRANS_PER_DOC: number;
    MAX_TRANS_PER_SUPP: number;
    MAX_TRANS_PER_CUST: number;
    SUPP_MTCH_COUNT: number;
    SUPP_MTCH_PERCENT: number;
    AV_SUPP_COUNT: number;
    AV_PROSPECT_COUNT: number;
    AV_SUPP_PERCENT: number;
    AV_PROSPECT_PERCENT: number;
    NEW_SUPP_MTCH_COUNT: number;
    CUST_MTCH_COUNT: number;
    CUST_MTCH_PERCENT: number;
    NEW_CUST_MTCH_COUNT: number;
    SPEND_TOTAL: number;
    SPEND_IN_SCOPE: number;
    SPEND_OUT_SCOPE: number;
    SPEND_PERCENT_IN_SCOPE: number;
    SPEND_PERCENT_OUT_SCOPE: number;
    GOOD_SPEND_AMT: number;
    BAD_SPEND_AMT: number;
    GOOD_SPEND_PERCENT: number;
    BAD_SPEND_PERCENT: number;
    TRANS_EXCLUDED_COUNT: number;
    TRANS_EXCLUDED_LINE_COUNT: number;
    TRANS_EXCLUDED_AMT: number;
    QA_REVIEW_STATUS: string;
    QA_REVIEW_ACTION_CODE: string;
    QA_REVIEW_ACTION_DESCR: string;
    LAST_REVIEWED_BY: string;
    QA_PASS_STATUS: string;
    QA_REVIEW_COMPLETE_DATE: Date;
    QA_PASSED_DATE: Date;
    QA_PASSED_BY: string;
    PROCESS_ID: number;
    CREATED_BY: string;
    CREATED_DATE: Date;
}